add room support in window tab

+++ app/src/main/java/com/faircorp/model/WindowDto.kt
@@ -2,5 +2,5 @@ package com.faircorp.model
 
 enum class windowStatus { OPEN, CLOSED}
 
-data class WindowDto(val id: Long, val name: String, val windowStatus: windowStatus, val roomName: String, val roomId: Int)
+data class WindowDto(val id: Long, val name: String, val windowStatus: windowStatus, val roomName: String, val roomId: Long)
